requirements-dev.txt
# Development Requirements for SYSGrow Backend
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# ===== DEVELOPMENT TOOLS =====
# Code formatting and linting
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
pylint>=2.17.0
mypy>=1.5.0

# ===== TESTING FRAMEWORK =====
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # For parallel test execution
coverage>=7.3.0

# ===== DOCUMENTATION =====
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# ===== DEBUGGING & PROFILING =====
ipdb>=0.13.0
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.0

# ===== CODE QUALITY =====
pre-commit>=3.4.0
bandit>=1.7.0  # Security linting
safety>=2.3.0  # Check for known security vulnerabilities

# ===== JUPYTER NOTEBOOKS =====
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
nbconvert>=7.7.0

# ===== DATABASE TOOLS =====
sqlite-utils>=3.34.0
sqlalchemy>=2.0.0  # For advanced database operations

# ===== API DEVELOPMENT =====
httpie>=3.2.0
postman-cli>=1.0.0

# ===== ENVIRONMENT MANAGEMENT =====
python-dotenv>=1.0.0
environs>=9.5.0

# ===== MOCK SERVICES =====
responses>=0.23.0
requests-mock>=1.11.0
fakeredis>=2.18.0